Chore: Permit installation on Python 3.7 and higher
Rationale: While Python 3.7 is already end-of-life, some people still
need to support it. Because `click-aliases` is often pulled into
projects using click, it is very common, and as such, currently imposes
a restriction which might not exactly be necessary.

It would be sad needing to vendor the code only to support Python 3.7
for people who need it.